fix(observability): accept typed attribute plugin mappings
