+++ crates/core/src/observability/plugin_component.rs
@@ -1563,6 +1563,7 @@ fn validate_observability_section_fields(
             "enabled",
             "mark_projection",
             "mark_exclude_names",
+            "attribute_mappings",
             "transport",
             "endpoint",
             "headers",
@@ -1583,6 +1584,7 @@ fn validate_observability_section_fields(
             "enabled",
             "mark_projection",
             "mark_exclude_names",
+            "attribute_mappings",
             "transport",
             "endpoint",
             "headers",
